Separate default game library
