+++ SlashGaming-Diablo-II-API/src/cxx/default_game_library.cc
@@ -139,62 +139,3 @@ GetDefaultLibraryPathWithRedirect(
 }
 
 } // namespace sgd2mapi
-
-char*
-SGD2MAPI_GetGameExecutablePath(
-  char dest[]
-) {
-  std::string game_executable_path_text =
-      sgd2mapi::GetGameExecutablePath().u8string();
-
-  std::copy(
-      game_executable_path_text.cbegin(),
-      game_executable_path_text.cend(),
-      dest
-  );
-
-  return dest;
-}
-
-std::size_t
-SGD2MAPI_GetGameExecutablePathSize(
-    void
-) {
-  return sgd2mapi::GetGameExecutablePath().u8string().size() + 1;
-}
-
-char*
-GetDefaultLibraryPathWithRedirect(
-    char dest[],
-    enum SGD2MAPI_DefaultLibrary library
-) {
-  enum sgd2mapi::DefaultLibrary actual_default_library =
-      static_cast<sgd2mapi::DefaultLibrary>(library);
-
-  const std::filesystem::path& default_library_path =
-      sgd2mapi::GetDefaultLibraryPathWithRedirect(actual_default_library);
-
-  std::string default_library_path_text =
-      default_library_path.u8string();
-
-  std::copy(
-      default_library_path_text.cbegin(),
-      default_library_path_text.cend(),
-      dest
-  );
-
-  return dest;
-}
-
-std::size_t
-GetDefaultLibraryPathSizeWithRedirect(
-    enum SGD2MAPI_DefaultLibrary library
-) {
-  enum sgd2mapi::DefaultLibrary actual_default_library =
-      static_cast<sgd2mapi::DefaultLibrary>(library);
-
-  const std::filesystem::path& default_library_path =
-      sgd2mapi::GetDefaultLibraryPathWithRedirect(actual_default_library);
-
-  return default_library_path.u8string().size() + 1;
-}
